made inventory list reusable
