+++ app/Components/TreeInventory/index.js
@@ -7,7 +7,7 @@ import {
   ScrollView,
   ActivityIndicator,
 } from 'react-native';
-import { Header, SmallHeader, InventoryCard, PrimaryButton } from '../Common';
+import { Header, SmallHeader, InventoryCard, PrimaryButton, InventoryList } from '../Common';
 import { SafeAreaView } from 'react-native';
 import { getAllInventoryByStatus, clearAllIncompleteInventory } from '../../Actions';
 import { store } from '../../Actions/store';
@@ -31,76 +31,10 @@ const TreeInventory = ({ navigation }) => {
     return unsubscribe;
   }, [navigation]);
 
-  const onPressInventory = (item) => {
-    setTimeout(() => {
-      dispatch(LocalInventoryActions.setInventoryId(item.inventory_id));
-      navigation.navigate(item.last_screen);
-    }, 0);
-  };
-
   const initialState = () => {
     getAllInventoryByStatus('all').then((allInventory) => {
       setAllInventory(Object.values(allInventory));
     });
-  };
-
-  const renderInventoryList = (inventoryList) => {
-    return (
-      <FlatList
-        showsVerticalScrollIndicator={false}
-        data={inventoryList}
-        renderItem={({ item }) => {
-          let imageURL;
-          let isOffSitePoint = false;
-          if (
-            item.polygons[0] &&
-            item.polygons[0].coordinates &&
-            Object.values(item.polygons[0].coordinates).length
-          ) {
-            imageURL = item.polygons[0].coordinates[0].imageUrl;
-            isOffSitePoint = Object.values(item.polygons[0].coordinates).length == 1;
-          }
-          let locateTreeAndType = '';
-          let title = '';
-          if (item.locate_tree === 'off-site') {
-            locateTreeAndType = i18next.t('label.tree_inventory_off_site');
-          } else {
-            locateTreeAndType = i18next.t('label.tree_inventory_on_site');
-          }
-          if (item.tree_type == 'single') {
-            title = `1 ${item.specei_name ? `${item.specei_name} ` : ''}` + i18next.t('label.tree_inventory_tree');
-            locateTreeAndType += ' - ' + i18next.t('label.tree_inventory_point');
-          } else {
-            let totalTreeCount = 0;
-            let species = Object.values(item.species);
-
-            for (let i = 0; i < species.length; i++) {
-              const oneSpecies = species[i];
-              totalTreeCount += Number(oneSpecies.treeCount);
-            }
-            title = `${totalTreeCount} ` + i18next.t('label.tree_inventory_trees');
-            locateTreeAndType += ` - ${isOffSitePoint ? i18next.t('label.tree_inventory_point') : i18next.t('label.tree_inventory_polygon')}`;
-          }
-          let data = {
-            title: title,
-            subHeading: locateTreeAndType,
-            date: i18next.t('label.tree_inventory_overview_date', {
-              date: new Date(Number(item.plantation_date)),
-            }),
-            imageURL: imageURL,
-          };
-          return (
-            <TouchableOpacity
-              onPress={() => onPressInventory(item)}
-              accessibilityLabel={i18next.t('label.tree_inventory_inventory_list')}
-              accessible={true}
-              testID="inventory_list">
-              <InventoryCard icon={'cloud-outline'} data={data} />
-            </TouchableOpacity>
-          );
-        }}
-      />
-    );
   };
 
   const onPressClearAll = () => {
@@ -143,7 +77,10 @@ const TreeInventory = ({ navigation }) => {
               icon={'cloud-upload'}
               style={{ marginVertical: 15 }}
             />
-            {renderInventoryList(pendingInventory)}
+            <InventoryList
+              accessibilityLabel={i18next.t('label.tree_inventory_inventory_list')}
+              inventoryList={pendingInventory}
+            />
           </>
         )}
         {uploadedInventory.length > 0 && (
@@ -163,7 +100,10 @@ const TreeInventory = ({ navigation }) => {
               rightTheme={'red'}
               style={{ marginVertical: 15 }}
             />
-            {renderInventoryList(inCompleteInventory)}
+            <InventoryList
+              accessibilityLabel={i18next.t('label.tree_inventory_inventory_list')}
+              inventoryList={inCompleteInventory}
+            />
           </>
         )}
       </View>
