+++ app/Components/UploadedInventory/index.js
@@ -3,12 +3,11 @@ import {
   View,
   StyleSheet,
   TouchableOpacity,
-  FlatList,
   ScrollView,
   ActivityIndicator,
   Text,
 } from 'react-native';
-import { Header, InventoryCard, PrimaryButton, AlertModal } from '../Common';
+import { Header, InventoryList, PrimaryButton, AlertModal } from '../Common';
 import { SafeAreaView } from 'react-native';
 import { getAllInventoryByStatus, clearAllUploadedInventory } from '../../Actions';
 import { store } from '../../Actions/store';
@@ -38,13 +37,6 @@ const UploadedInventory = ({ navigation }) => {
     });
   };
 
-  const onPressInventory = (item) => {
-    setTimeout(() => {
-      dispatch(LocalInventoryActions.setInventoryId(item.inventory_id));
-      navigation.navigate(item.last_screen);
-    }, 0);
-  };
-
   const freeUpSpace = () => {
     clearAllUploadedInventory().then(() => {
       initialState();
@@ -54,66 +46,6 @@ const UploadedInventory = ({ navigation }) => {
 
   const toogleIsShowFreeUpSpaceAlert = () => {
     setIsShowFreeUpSpaceAlert(!isShowFreeUpSpaceAlert);
-  };
-
-  const renderInventoryList = (inventoryList) => {
-    return (
-      <FlatList
-        showsVerticalScrollIndicator={false}
-        data={inventoryList}
-        renderItem={({ item }) => {
-          let imageURL;
-          let isOffSitePoint = false;
-          if (
-            item.polygons[0] &&
-            item.polygons[0].coordinates &&
-            Object.values(item.polygons[0].coordinates).length
-          ) {
-            imageURL = item.polygons[0].coordinates[0].imageUrl;
-            isOffSitePoint = Object.values(item.polygons[0].coordinates).length == 1;
-          }
-          let locateTreeAndType = '';
-          let title = '';
-          if (item.locate_tree === 'off-site') {
-            locateTreeAndType = i18next.t('label.tree_inventory_off_site');
-          } else {
-            locateTreeAndType = i18next.t('label.tree_inventory_on_site');
-          }
-          if (item.tree_type == 'single') {
-            title = `1 ${item.specei_name ? `${item.specei_name} ` : ''}` + i18next.t('label.tree_inventory_tree');
-            locateTreeAndType += ' - ' + i18next.t('label.tree_inventory_point');
-          } else {
-            let totalTreeCount = 0;
-            let species = Object.values(item.species);
-
-            for (let i = 0; i < species.length; i++) {
-              const oneSpecies = species[i];
-              totalTreeCount += Number(oneSpecies.treeCount);
-            }
-            title = `${totalTreeCount} ` + i18next.t('label.tree_inventory_trees');
-            locateTreeAndType += ` - ${isOffSitePoint ? i18next.t('label.tree_inventory_point') : i18next.t('label.tree_inventory_polygon')}`;
-          }
-          let data = {
-            title: title,
-            subHeading: locateTreeAndType,
-            date: i18next.t('label.inventory_overview_date', {
-              date: new Date(Number(item.plantation_date)),
-            }),
-            imageURL: imageURL,
-          };
-
-          return (
-            <TouchableOpacity
-              onPress={() => onPressInventory(item)}
-              accessible={true}
-              accessibilityLabel={i18next.t('label.tree_inventory_upload_inventory_list')}
-              testID="upload_inventory_list">
-              <InventoryCard icon={'cloud-check'} data={data} />
-            </TouchableOpacity>
-          );
-        }}
-      />
-    );
   };
 
   const renderInventory = () => {
@@ -126,9 +58,14 @@ const UploadedInventory = ({ navigation }) => {
               accessible={true}
               accessibilityLabel={i18next.t('label.tree_inventory_free_up_space')}
               testID="free_up_space">
-              <Text style={styles.freeUpSpace}>{i18next.t('label.tree_inventory_free_up_space')}</Text>
+              <Text style={styles.freeUpSpace}>
+                {i18next.t('label.tree_inventory_free_up_space')}
+              </Text>
             </TouchableOpacity>
-            {renderInventoryList(allInventory)}
+            <InventoryList
+              accessibilityLabel={i18next.t('label.tree_inventory_upload_inventory_list')}
+              inventoryList={allInventory}
+            />
           </>
         )}
       </View>
